ai-service/requirements.txt
fastapi
uvicorn
langchain>=0.3
langchain-core>=0.3
langchain-google-genai>=2.1
langchain-community>=0.3
langchain-openai>=0.3
python-dotenv
psycopg2-binary
pydantic
